requirements-docker.txt
# Requirements for Docker migration
pandas==2.1.4
openpyxl==3.1.2
mysql-connector-python==8.2.0
python-dotenv==1.0.0

